% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloverblik_data_access_token.R
\name{download_data_access_token}
\alias{download_data_access_token}
\title{Download data access token for Eloverblik}
\usage{
download_data_access_token(refresh_token = get_refresh_token())
}
\arguments{
\item{refresh_token}{A valid refresh token}
}
\value{
A \code{{httr}} response.
}
\description{
Download data access token for Eloverblik
}
