% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energidataservice_spot_prices.R
\name{download_spot_prices}
\alias{download_spot_prices}
\title{Download spot price data}
\usage{
download_spot_prices(start, end, area)
}
\arguments{
\item{start}{\verb{[Date]} The start date.}

\item{end}{\verb{[Date]} The end date.}

\item{area}{\verb{[character(1)]} The area in Denmark. Allowed values are "DK1" and "DK2".}
}
\value{
A \code{{httr}} response.
}
\description{
The API is described here: \url{https://www.energidataservice.dk/guides/api-guides}.
}
