% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloverblik_data_access_token.R
\name{get_data_access_token}
\alias{get_data_access_token}
\title{Get data access token for Eloverblik}
\usage{
get_data_access_token(refresh_token = get_refresh_token())
}
\arguments{
\item{refresh_token}{A valid refresh token}
}
\value{
The data access token as a string.
}
\description{
Get data access token for Eloverblik
}
