% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloverblik_refresh_token.R
\name{get_refresh_token}
\alias{get_refresh_token}
\title{Retrieve refresh token}
\usage{
get_refresh_token()
}
\description{
Read refresh token from the environment variable \code{ELOVERBLIK_TOKEN}, if available.
}
