% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{lazy_daily_prices}
\alias{lazy_daily_prices}
\title{Compute daily consumption prices}
\usage{
lazy_daily_prices(price_tbl)
}
\arguments{
\item{price_tbl}{Output from \code{\link[=lazy_read_prices]{lazy_read_prices()}}.}
}
\value{
An \code{\link[arrow:Dataset]{arrow::FileSystemDataset()}}.
}
\description{
Compute daily consumption prices
}
