% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{lazy_join_all}
\alias{lazy_join_all}
\title{Join consumption data and price data}
\usage{
lazy_join_all(meter_data, spot_prices, fees)
}
\arguments{
\item{meter_data}{Output from \code{\link[=lazy_read_meter_data]{lazy_read_meter_data()}}.}

\item{spot_prices}{Output from \code{\link[=lazy_read_spot_prices]{lazy_read_spot_prices()}}.}

\item{fees}{Output from \code{\link[=lazy_read_fees]{lazy_read_fees()}}.}
}
\value{
An \code{\link[arrow:Dataset]{arrow::FileSystemDataset()}}.
}
\description{
Lazy join of consumption data and price data.
}
