% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{lazy_read_fees}
\alias{lazy_read_fees}
\title{Lazy load of fee data}
\usage{
lazy_read_fees(fees_folder = here::here("data", "fees"), margin = 0.1)
}
\arguments{
\item{spot_price_folder}{Path to the root of the fee data.}
}
\value{
An \code{\link[arrow:Table]{arrow::Table()}}.
}
\description{
There are a number of function to \strong{lazily} load all data used in {eldata}. Check the details
for more info about lazy data.
}
\details{
Lazy load means that we return a query from the \code{{arrow}} package, but no \strong{data} is
available in R. This defers computations and (more importantly) loading of data into R so we only
read the necessary data. Use \code{\link[dplyr:compute]{dplyr::collect()}} to read the data into R.

Check these resources for more info about the \code{{arrow}} package:
\itemize{
\item The "dataset" vignette in the arrow package: \code{vignette('dataset', package = 'arrow')}.
\item \url{https://arrow-user2022.netlify.app/data-wrangling.html}
\item \url{https://arrow.apache.org/docs/r//reference/acero.html}
}
}
