% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{lazy_read_meter_data}
\alias{lazy_read_meter_data}
\title{Lazy load of meter data}
\usage{
lazy_read_meter_data(
  meter_data_folder = here::here("data", "parsed", "meter_data")
)
}
\arguments{
\item{meter_data_folder}{Path to the root of the meter data.}
}
\value{
An \code{\link[arrow:Dataset]{arrow::FileSystemDataset()}}.
}
\description{
There are a number of function to \strong{lazily} load all data used in {eldata}. Check the details
for more info about lazy data.
}
\details{
Lazy load means that we return a query from the \code{{arrow}} package, but no \strong{data} is
available in R. This defers computations and (more importantly) loading of data into R so we only
read the necessary data. Use \code{\link[dplyr:compute]{dplyr::collect()}} to read the data into R.

Check these resources for more info about the \code{{arrow}} package:
\itemize{
\item The "dataset" vignette in the arrow package: \code{vignette('dataset', package = 'arrow')}.
\item \url{https://arrow-user2022.netlify.app/data-wrangling.html}
\item \url{https://arrow.apache.org/docs/r//reference/acero.html}
}
}
