% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{lazy_read_prices}
\alias{lazy_read_prices}
\title{Compute hourly consumption prices}
\usage{
lazy_read_prices(consumption_and_prices)
}
\arguments{
\item{consumption_and_prices}{Output from \code{\link[=lazy_join_all]{lazy_join_all()}}}
}
\value{
An \code{\link[arrow:Dataset]{arrow::FileSystemDataset()}}.
}
\description{
Compute hourly consumption prices
}
