% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloverblik_meter_data.R
\name{load_meter_data}
\alias{load_meter_data}
\title{Load saved meter data}
\usage{
load_meter_data(raw_folder)
}
\arguments{
\item{raw_folder}{Folder with JSON files.}
}
\value{
A \code{tibble} with columns \code{Date}, \code{HourOfDay}, \code{StartTimeUTC}, \code{EndTimeUTC}, \code{Consumption}, \code{Quality}, \code{Resolution}.
}
\description{
Raw meter data saved in JSON is loaded and parsed.
}
