% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloverblik_data_access_token.R
\name{parse_data_access_token}
\alias{parse_data_access_token}
\title{Parse data access token response}
\usage{
parse_data_access_token(response)
}
\arguments{
\item{response}{Response from \link{download_data_access_token}.}
}
\description{
Parse data access token response
}
