% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloverblik_meter_data.R
\name{parse_meter_data}
\alias{parse_meter_data}
\title{Parse meter data response}
\usage{
parse_meter_data(json_file)
}
\arguments{
\item{json_file}{JSON content from a Response from \link{download_spot_prices}.}
}
\value{
A \code{tibble} with columns \code{Date}, \code{HourOfDay}, \code{StartTimeUTC}, \code{EndTimeUTC}, \code{Consumption}, \code{Quality}, \code{Resolution}.
}
\description{
Parse meter data into a \code{tibble}.
}
\details{
Because the data source is brittle and I only include a subset of the data in the output of this function, I save the raw JSON content first instead of parsing the response.
}
