% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_metering_points_info}
\alias{read_metering_points_info}
\title{Read meter points info}
\usage{
read_metering_points_info(
  metering_points_file = here::here("data", "metering_points.csv"),
  delim = ";"
)
}
\arguments{
\item{metering_points_file}{The location of the CSV file.}

\item{delim}{The delimiter passed on to \code{\link[readr:read_delim]{readr::read_delim()}}.}
}
\value{
A tibble.
}
\description{
Metering points info in CSV format is expected to have columns:
}
\details{
\itemize{
\item \code{MeterId}
\item \code{TokenId}
\item \code{Name}
}
}
