% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloverblik_meter_data.R
\name{save_parsed_meter_data}
\alias{save_parsed_meter_data}
\title{Save parsed meter data}
\usage{
save_parsed_meter_data(meter_data, parsed_folder)
}
\arguments{
\item{meter_data}{Dataframe from \link{parse_meter_data}.}

\item{parsed_folder}{Folder to store CSV files.}
}
\description{
Save parsed meter data to CSV files -- one file per metering id.
}
