% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloverblik_meter_data.R
\name{download_meter_data}
\alias{download_meter_data}
\title{Download meter data}
\usage{
download_meter_data(dateFrom, dateTo, aggregation, meteringpoint, dat)
}
\arguments{
\item{dateFrom}{\verb{[Date]} The start date.}

\item{dateTo}{\verb{[Date]} The end date.}

\item{aggregation}{\verb{[character(1)]} Time aggregation. Allowed values are "Actual", "Quarter", "Hour", "Day", "Month", "Year".}

\item{meteringpoint}{\verb{[character(1)]} A single metering point to get data for.}

\item{dat}{\verb{[character(1)]} Data access token from \code{\link[=get_data_access_token]{get_data_access_token()}}.}
}
\value{
A \code{{httr}} response.
}
\description{
The API is described on a Swagger page: \url{https://api.eloverblik.dk/CustomerApi/index.html}.
}
